\name{dcglm-package}
\alias{dcglm-package}
\alias{dcglm}
\docType{package}
\title{
What the package does (short line)
Tutorial package for writing packages based on data cloning
}
\description{
Tutorial package for writing packages based on data cloning
}
\author{
Peter Solymos \email{solymos@ulaberta.ca}
}
\references{
Lele, S.R., B. Dennis and F. Lutscher, 2007.
Data cloning: easy maximum likelihood estimation for complex 
ecological models using Bayesian Markov chain Monte Carlo  methods.
\emph{Ecology Letters} \strong{10}, 551--563.
}
\keyword{ package }
\examples{
## see vignette
}
